Split useInvitation into two
