+++ packages/react/src/index.ts
@@ -1,10 +1,11 @@
+import useAcceptInvitation from './useAcceptInvitation'
 import useAcceptOffer, { AcceptOfferStep } from './useAcceptOffer'
 import useAuthenticate from './useAuthenticate'
 import useBatchPurchase, { BatchPurchaseStep } from './useBatchPurchase'
 import useCancelOffer, { CancelOfferStep } from './useCancelOffer'
+import useCreateInvitation from './useCreateInvitation'
 import useCreateNFT, { CreateNftStep } from './useCreateNFT'
 import useCreateOffer, { CreateOfferStep } from './useCreateOffer'
-import useInvitation from './useInvitation'
 import useIsLoggedIn from './useIsLoggedIn'
 import useMintDrop, { MintDropStep } from './useMintDrop'
 import useUpdateAccount from './useUpdateAccount'
@@ -29,7 +30,12 @@ export {
 export { CreateNftStep, MintDropStep, useCreateNFT, useMintDrop }
 
 // Account
-export { useInvitation, useUpdateAccount, useVerifyAccount }
+export {
+  useAcceptInvitation,
+  useCreateInvitation,
+  useUpdateAccount,
+  useVerifyAccount,
+}
 
 // Auth
 export { useAuthenticate, useIsLoggedIn }
